Add files via upload
This is another project I worked on. It honestly took me one whole day to understand how CSS-Grid actually works, due to the fact I had a lot of nested Div within the parent container. I did this purposely to understand how the grid system works.

Also I added a comment section below the site with JavaScript with a little improvement with the Navbar.